extract downloading strategy
