+++ crates/stages/src/stages/headers/mod.rs
@@ -1,0 +1,5 @@
+/// The headers stage implementation
+pub mod stage;
+
+/// The downloading strategies
+pub mod linear;
